refactored and moved to director-v2
